refactor(Tracker): move Jira into its own folder
This opens up the possibility of having more trackers

For that to happen config would need to be refactored though
