+++ src/commands/Review.ts
@@ -5,7 +5,7 @@ import { Git } from 'src/git/Git';
 import { Github } from 'src/git/Github';
 import { PullRequest, Remote } from 'src/git/Remote';
 import { Emoji, Messenger } from 'src/io/messenger';
-import { Jira } from 'src/issue-tracker/Jira';
+import { Jira } from 'src/issue-tracker/jira/Jira';
 import { Tracker } from 'src/issue-tracker/Tracker';
 import { Issue, IssueStatus } from 'src/types';
