+++ src/commands/Start.ts
@@ -4,7 +4,7 @@ import { map, switchMap, tap, withLatestFrom } from 'rxjs/operators';
 import { FotingoArguments } from 'src/commands/FotingoArguments';
 import { Git } from 'src/git/Git';
 import { Emoji, Messenger } from 'src/io/messenger';
-import { Jira } from 'src/issue-tracker/Jira';
+import { Jira } from 'src/issue-tracker/jira/Jira';
 import { Tracker } from 'src/issue-tracker/Tracker';
 import { CreateIssue, GetIssue, Issue, IssueStatus, IssueType } from 'src/types';
