+++ src/config.ts
@@ -8,7 +8,7 @@ import * as path from 'path';
 import * as R from 'ramda';
 
 import { GitConfig, GithubConfig } from './git/Config';
-import { JiraConfig } from './issue-tracker/Config';
+import { JiraConfig } from './issue-tracker/jira/Config';
 import { ReleaseConfig } from './types';
 
 export interface Config {
