+++ src/issue-tracker/jira/JiraError.ts
